Code cleanup and organized
