+++ jfd-backend/src/main/java/com/example/controller/HomeController.java
@@ -1,0 +1,29 @@
+package com.example.controller;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.example.entity.User;
+import com.example.repository.UserRepository;
+
+@RestController
+public class HomeController {
+
+	@Autowired
+	private UserRepository userRepository;
+	
+	@GetMapping("/users")
+	public List<User> users() {
+		Iterable<User> users = userRepository.findAll();
+		ArrayList<User> usersList = new ArrayList<User>();
+		for (User user : users) {
+			usersList.add(user);
+		}
+
+		return usersList;
+	}
+}
