+++ jfd-frontend/src/app/service/signup.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { RegisterService } from './register.service';
+import { SignupService } from './signup.service';
 
-describe('RegisterService', () => {
-  let service: RegisterService;
+describe('SignupService', () => {
+  let service: SignupService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(RegisterService);
+    service = TestBed.inject(SignupService);
   });
 
   it('should be created', () => {
