--- jfd-frontend/src/app/signup/greeting.model.ts
@@ -1,4 +1,0 @@
-export interface greeting {
-    id: number;
-    content: string;
-}
